+++ areas/storage/src/AzureMcp.Storage/Options/Queue/Message/QueueMessageSendOptions.cs
@@ -1,0 +1,18 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+
+namespace AzureMcp.Storage.Options.Queue.Message;
+
+public class QueueMessageSendOptions : BaseQueueOptions
+{
+    [JsonPropertyName(StorageOptionDefinitions.MessageContent)]
+    public string? MessageContent { get; set; }
+
+    [JsonPropertyName(StorageOptionDefinitions.TimeToLiveInSeconds)]
+    public int? TimeToLiveInSeconds { get; set; }
+
+    [JsonPropertyName(StorageOptionDefinitions.VisibilityTimeoutInSeconds)]
+    public int? VisibilityTimeoutInSeconds { get; set; }
+}